+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master
 - saving a new session does not remove the previous one
+- make the directory where sessions are stored configurable
 
 ### v0.0.1, 2014-08-26
 - started a project